requirements.txt
# CloudOps-SRE-Toolkit Requirements

# Core dependencies
requests>=2.31.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0

# Cloud SDKs
boto3>=1.26.0
azure-mgmt-costmanagement>=3.0.0
azure-mgmt-resource>=22.0.0
azure-identity>=1.12.0
google-cloud-billing>=1.10.0

# Kubernetes
kubernetes>=27.2.0

# Security and SSL
cryptography>=41.0.0
pyOpenSSL>=23.2.0

# Monitoring and observability
aiohttp>=3.8.0
prometheus-client>=0.16.0

# Data processing
jq>=1.6.0  # Note: This is a Python wrapper for jq command

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Optional dependencies for enhanced functionality
# Uncomment as needed:

# For advanced logging
# structlog>=23.1.0

# For configuration management
# pydantic>=2.1.0

# For async operations
# asyncio-mqtt>=0.13.0

# For database operations
# sqlalchemy>=2.0.0

# For time series analysis
# scipy>=1.11.0

# For machine learning (future features)
# scikit-learn>=1.3.0
